Update User endpoints and services to support authenticated logins and verifications

+++ app/services/user_service.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm import Session
 from fastapi import HTTPException
 from app.models.user import User
 from app.schemas import UserCreate
+from app.utils.security import get_pwd_hash
 
 def get_user(user_id:int , db:Session):
     user = db.query(User).filter(User.id == user_id).first()
@@ -19,12 +20,15 @@ def create_user(user:UserCreate, db:Session):
     if db.query(User).filter(User.email == user.email).first():
         raise HTTPException(status_code=400, detail="User already exists!")
     
+    hashed_password = get_pwd_hash(user.password)
+    
     # Create new user
     new_user = User(name=user.name,
                    age=user.age,
                    location=user.location, 
                    email = user.email, 
-                   role=user.role)
+                   role=user.role,
+                   hashed_pswd=hashed_password)
 
     db.add(new_user)
     db.commit()
@@ -40,6 +44,7 @@ def update_user(user_id:int, user:UserCreate, db:Session):
     # Update User
     db_user.name = user.name
     db_user.age = user.age
+    db_user.location = user.location
     db_user.email = user.email
     db_user.role = user.role
 
@@ -47,11 +52,14 @@ def update_user(user_id:int, user:UserCreate, db:Session):
     db.refresh(db_user)
     return db_user
 
-def delete_user(user_id:int, db:Session):
+def delete_user(user_id:int, current_user, db:Session):
     user = db.query(User).filter(User.id == user_id).first()
     
     if not user:
         raise HTTPException(status_code=404, detail="User doesn't exist")
+    
+    if user.id == current_user.id:
+        raise HTTPException(status_code=404, detail="You can't delete yourself!")
     
     db.delete(user)
     db.commit()
